Separate curr_timestamp shared by scheduler and recv_callback
Previously, scheduler and recv_callback share the same variable for storing
the current timestamp. However, it might be possible that the recv_callback
might be pre-empted by the scheduler and cause the curr_timestamp variable
to be overwritten by the scheduler. This might cause the curr_timestamp in
used by recv_callback to become inaccurate.

Out of abundance of caution, let's separate the curr_timestamp variable used
by the scheduler and recv_callback, and make the scheduler use its own variable.