Remove nodes that are in ABSENT but registered
Previously, we only remove registered and DETECT-ed nodes.

However, nodes that are registered but fail to reach the DETECT stage
will not be cleared even if it has died.

Let's make it such that we clear the node from the memory map if the time
difference between last detection is greater than the out of proximity
threshold, irregardless of their state.